security fix!

+++ trunk/client/DirectoryListing.cpp
@@ -106,7 +106,7 @@ void DirectoryListing::loadFile(const string& name) {
 		}
 	} else if(Util::stricmp(ext, ".xml") == 0) {
 		int64_t sz = ::File::getSize(name);
-		if(sz == -1 || sz >= txt.max_size())
+		if(sz == -1 || sz >= static_cast<int64_t>(txt.max_size()))
 			throw(FileException(CSTRING(FILE_NOT_AVAILABLE)));
 		txt.resize((size_t) sz);
 		size_t n = txt.length();
